VRP_new_app/requirements.txt
streamlit
pandas
requests
geopy
ortools
folium
polyline
numpy
tqdm
playwright  # Added for E2E testing
pytest-playwright  # Added for E2E testing
tenacity==8.2.3
